backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
pytesseract
Pillow
sentence-transformers
chromadb
langchain
langchain-community
openai
